rename and refactor event to report
