+++ src/components/ReportForm/ReportForm.css
@@ -1,4 +1,4 @@
-.AddEvent {
+.ReportForm {
   display: flex;
   position: relative;
   height: 100%;
